+++ src/app/app.module.ts
@@ -2,11 +2,10 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { CoreModule } from './core/core.module';
 import { AppRoutingModule } from './core/app-routing.module';
-import { FormsModule } from '@angular/forms';
 import { HomeModule } from './home/home.module';
 
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {MatIconModule, MatButtonModule, MatCardModule, MatDialogModule, MatSelectModule} from '@angular/material';
+import {MatIconModule, MatButtonModule, MatCardModule} from '@angular/material';
 
 import { AppComponent } from './app.component';
 import { NavComponent } from './nav/nav.component';
@@ -22,7 +21,6 @@ import { AngularFirestoreModule } from 'angularfire2/firestore';
 import { DatabaseService } from './database.service';
 import { WelcomeComponent } from './welcome/welcome.component';
 import { AdminComponent } from './admin/admin/admin.component';
-import { TransactionsComponent } from './transactions/transactions.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 
 export const firebaseConfig = environment.firebaseConfig;
@@ -33,7 +31,6 @@ export const firebaseConfig = environment.firebaseConfig;
     NavComponent,
     WelcomeComponent,
     AdminComponent,
-    TransactionsComponent,
     PageNotFoundComponent
   ],
   imports: [
@@ -43,9 +40,6 @@ export const firebaseConfig = environment.firebaseConfig;
     MatButtonModule,
     BrowserAnimationsModule,
     MatCardModule,
-    MatDialogModule,
-    MatSelectModule,
-    FormsModule,
     AngularFireModule.initializeApp(firebaseConfig),
     AngularFireAuthModule,
     CoreModule,
@@ -54,6 +48,5 @@ export const firebaseConfig = environment.firebaseConfig;
   ],
   providers: [PosService, DatabaseService],
   bootstrap: [AppComponent],
-  entryComponents: []
 })
 export class AppModule { }
