+++ src/app/home/history/history.component.html
@@ -10,5 +10,8 @@
     <mat-divider></mat-divider>
   </mat-list>
 
+  <mat-card-actions>
+    <button mat-button routerLink="/transactions">View All Transactions</button>
+  </mat-card-actions>
 
 </mat-card>
