+++ src/app/home/home.module.ts
@@ -1,13 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { AppRoutingModule } from '../core/app-routing.module';
+import { FormsModule } from '@angular/forms';
+
 
 import { HistoryComponent } from './history/history.component';
 import { PosComponent } from './pos/pos.component';
 import { TicketComponent } from './ticket/ticket.component';
 import { HomeComponent } from './home.component';
 import { LineItemModalComponent } from './history/line-item-modal/line-item-modal.component';
+import { TransactionsComponent } from './transactions/transactions.component';
 
-import {MatIconModule, MatButtonModule, MatTabsModule, MatCardModule, MatDialogModule} from '@angular/material';
+import {MatIconModule, MatButtonModule, MatTabsModule, MatCardModule, MatDialogModule, MatSelectModule} from '@angular/material';
 
 
 @NgModule({
@@ -17,14 +21,18 @@ import {MatIconModule, MatButtonModule, MatTabsModule, MatCardModule, MatDialogM
     MatIconModule,
     MatButtonModule,
     MatDialogModule,
-    MatTabsModule
+    MatTabsModule,
+    MatSelectModule,
+    AppRoutingModule,
+    FormsModule
   ],
   declarations: [
     HomeComponent,
     TicketComponent,
     LineItemModalComponent,
     PosComponent,
-    HistoryComponent
+    HistoryComponent,
+    TransactionsComponent
   ],
   entryComponents: [LineItemModalComponent]
 })
