+++ src/app/home/transactions/transactions.component.scss
@@ -1,4 +1,4 @@
-@import '../app.component.scss';
+@import '../../app.component.scss';
 
 .container {
   margin: 1.5rem;
